Fixed RAP Downloading issue. Also removed unneccesary imports
