+++ src/EdgeWARN/PreProcess/CellDetection/tracker.py
@@ -1,7 +1,6 @@
 import json
 from .detect import detect_cells
 from pathlib import Path
-# assumes your existing imports: detect, load, match_cells, process_matched_cell, vectors, plot_radar_and_cells
 from ..core.utils import load_mrms_slice
 from ..core.save import StormCellTracker, save_cells_to_json
 from ..core.visualize import Visualizer
