+++ src/EdgeWARN/PreProcess/CellIntegration/integrator.py
@@ -1,4 +1,3 @@
-from util.core.file import StatFileHandler
 from .utils import StormIntegrationUtils
 from datetime import datetime
 from shapely.geometry import shape
