+++ src/EdgeWARN/PreProcess/core/cellmask.py
@@ -3,8 +3,7 @@ from scipy.ndimage import label, binary_dilation, center_of_mass
 import alphashape
 from shapely.geometry import Point, LineString
 import datetime
-from .utils import extract_timestamp_from_filename, CellProcessor, PENALTY_COST
-from scipy.optimize import linear_sum_assignment
+from .utils import extract_timestamp_from_filename
 
 class StormCellDetector:
     """
